[SILParser] work around PolymorphicFunctionType with input or result type being PolymorphicFunctionType.
SILParser needs to compare two PolymorphicFunctionType when resolving overloaded
SILDeclRef. SILParser currently performs a local canonicalization that does not
work with input or result type being PolymorphicFunctionType. This commit work
around that by assuming it is a match if both inputs are PolymorphicFunctionType
or both outputs are PolymorphicFunctionType.

rdar://18021608


Swift SVN r22147