Add AST pointer to TypeRefinementContext.
Updating r22145, this patch adds pointers from a TypeRefinementContext to the
AST node that introduced the context, along with the reason for refinement. This
reason supports the scenario where an AST node may introduce multiple refinement
contexts (e.g., eventually we want an IfStmt to introduce two refinement
contexts, one for the Then branch and one for the Else branch.

The patch also fixes a memory leak for the storage holding a refinement
context's children.


Swift SVN r22232