Add an option to swift-ide-test to test the debugger's handling of 'private'.
Specifically, to simulate the debugger's expression evaluation, where
declarations from a certain file are /preferred/ but not /required/.
This allows the debugger to reference private declarations unambiguously
based on its notion of the "current" source file.

This is the test case for r22034, which fixed rdar://problem/18296639.

Swift SVN r22271