Change the way CopyForwarding updates uses. Still disabled (NFC).
Make sure that we see a "deinit" of the copy and only replace uses in
the same block between the copy and the deinit. This should cover all
the cases I care about but is robust and no longer uses dominance.

While doing this cleanup, I added logic to handle Arnold's cases
where we need backward copy propagation.

This isn't super-pretty, still exposes bugs, and needs tests for the
cases that are actually working.

Swift SVN r22274