[Parser] generate conformances for substitutions in specialized conformance.
We share the implementation with generating conformances for substitutions in
ApplyInst. In addition to going through the requirements of the
GenericParamList, we also check the inherited list of the GenericParams.

rdar://18021102

With this commit, we can now parse stdlib with
bin/sil-opt -sil-disable-ast-dump -verify lib/swift/macosx/Swift.swiftmodule -module-name=Swift &> deser.sil
bin/sil-opt -verify deser.sil

But this test is time consuming, it runs for 2+ minutes on my machine, with
94% in performTypeChecking.


Swift SVN r22372