Don't first-pass type-check decls from other files when validation will do.
This corrects r11686 (!), which delayed the validation of members of a nominal
type until the end of type checking. The original commit used first-pass
type-checking to recursively check the decl and its members, but this does
more work than necessary---the original code just /validated/ the members.

The real purpose of this code is for layout and for vtable consistency, so
just validate all members and generate any implicit class members that may
be dynamically dispatched.

Swift SVN r22375