SILGen: Global accessors must always be non-fragile.
We can't expose the lazy initialization token across module boundaries because that might lead to it being duplicated, which would be bad, so if we were ever going to make them fragile, it would have to happen after a guaranteed globalopt pass. Fixes a crash that was blocking DaveA's work; his incoming patches should exercise this.

Swift SVN r22509