[Parser] parsing A<C, B<V>?>?.
This extends a workaround that handles A<C, B<V>?> to further handle
A<C, B<V>?>?. If we return ">?>?" as a single token, consumingStartingGreater
consumes ">", returns a token for "?" and silently drops the other characters.

The workaround is to return ">?" as a token when we see ">?>?".


Swift SVN r22585