Resolve regression resulting from enum_is_tag to select_enum transition.
Unwrap the argument to select_enum to determine if it is the same
condition as is used in the terminator we are looking to optimize.

This particular code has become a bit unwieldy and needs larger
generalization and refactoring, but this restores the previous
optimization for now.

Resolves rdar://problem/18592542.

Swift SVN r22655