Use the archetype builder’s type resolution more directly.
The type checker was building dependent types manually in a number of places where it would be better to let the archetype builder do it. Use the archetype builder, and start tracking those source locations where we have unresolved nested type names known to the archetype builder.

Swift SVN r22675