[Sema] Change diagnostic of useless #os() check to indicate why it is useless.
This commit updates diagnostic reporting for #os() checks that will always pass.
We now indicate whether the reason the check will always pass is because of the
minimum deployment target or is because of an enclosing scope. In the latter
case, we emit a note indicating the location of the enclosing scope.


Swift SVN r22729