SimplifyCFG: Split critical edges before updating SSA form
This way we are guaranteed to be able to build SSA form: there is guarantueed
to be a terminator branch that supports adding basic block arguments at every
join points. We were relying on the assumption that we will never have other
terminators at join points when we want to insert phi nodes. This turns out to
be a very fragile assumption in the presence of aggressive CFG simplification.

Swift SVN r22759