Add the flag DONT_STRIP_NON_MAIN_SYMBOLS to add_swift_executable.
Before this commit, we would always export only main from
executables. This implied that certain protocol conformances would be
stripped from the perf test suite. This includes an printable protocol
conformances. This created an issue since currently it is necessary to
use dlsym to get the witness table for that protocol. If we can't find
it due to stripping, we fall back to mirrors which are much slower and
cause certain of our tests (for example StringInterpolation) to test
mirror performance instead of what they are actually supposed to test.

*NOTE* By default, we will still use the relevant export list. This just
 adds an option to disable the export list.

rdar://18594587

Swift SVN r22805