[cmake] Add option boolean DISABLE_ASLR to add_swift_executable.
All this does is ensure that the linker is passed the flag -Wl,-no_pie
so that the final linked executable is not position independent. Since
the executable is not position independent it can not be relocated by
the kernel.

This is useful when one is attempting to debug hard to reproduce
miscompiles.

Swift SVN r22882