[stdlib] Use ''' instead of """ to deconfuse Emacs
The indenter does great until it sees the Python triple-quoted string
containing a swift string with a string intepolation inside; that brings
it out to quotation level zero and then the backslash escapes the open
paren causing the nesting level to go negative when it sees the close
paren.  Sheesh.

Swift SVN r22948