[func-sig-opts] Refactor func-sig-opts to make it easier to handle multiple @owned arguments.
The main change was to create a FunctionAnalyzer structure which stores
all of the data that we carry around. This reduces the number of
arguments for a bunch of functions and also enables the beginning of a
split in the pass in between the analysis part and the execution
part. There are still some things I need to clean up which I will do in
a little bit.

Swift SVN r22962