[test] Always update modules when checking PrintAsObjC output.
Previously if someone changed one of the mock SDK headers, it might not get
picked up by the PrintAsObjC tests because it treats those headers as
system headers. Swift doesn't have this problem because it always uses the
-fmodules-validate-system-headers option, but we should pass it explicitly
to Clang here for those not using clean builds / clean module caches.

Swift SVN r22982