[stdlib] Fix idiotic bug in ManagedBufferPointer
Found by inspection.  I tried long and hard to think of a test that
would detect this bug, and even tried to write quite a few, but none
would be triggered.  Bright ideas most welcome.

Commits to come, however, in the presence of this bug, cause around 25
failures in the stdlib test suite, so at least this won't regress.

Swift SVN r22984