Add test for transparent inlining with ObjC-protocol substitution.
We currently do not do mandatory inlining of generic functions in cases
where we would end up substituting existentials, but in the case of ObjC
protocols, we should be able to.

I'll fix that in a separate commit. We currently hit an assert if I
allow a SIL parsed (vs. Swift parsed) version of this test to be
inlined, and I need to investigate that first.

Swift SVN r23092